Refactor taskFormController
Remove the initialization of functions in the
constructor, instead make the functions not
private and move its initialization in the entry
file. Aside from this, closeForm method will be
removed since the modalController will handle it.
And change the code block in the showForm method.
Lastly, a new method,
changeSubmitButtonState will be added.